Move chain_header() for reuse
