+++ core/src/agent/mod.rs
@@ -3,3 +3,19 @@
 pub mod actions;
 pub mod chain_store;
 pub mod state;
+
+use context::Context;
+use holochain_core_types::{
+    cas::content::AddressableContent,
+    chain_header::ChainHeader,
+    entry::Entry,
+};
+use std::sync::Arc;
+
+pub fn chain_header(entry: &Entry, context: &Arc<Context>) -> Option<ChainHeader> {
+    let chain = context.state().unwrap().agent().chain();
+    let top_header = context.state().unwrap().agent().top_chain_header();
+    chain
+        .iter(&top_header)
+        .find(|ref header| *header.entry_address() == entry.address())
+}
